🔥 Reuse the Empty component
