+++ src/components/design/projects/index.tsx
@@ -1,9 +1,26 @@
 import * as React from 'react'
 import { IProject } from '../../../types'
-import NewProject from './components/new-project'
 import Header from './components/header'
 import List from './components/list'
 import { TeamsContext } from '../../../context/teamsContext'
+import { Empty } from '../empty'
+import { defineMessages } from 'react-intl'
+import { useHistory } from 'react-router-dom'
+
+const messages = defineMessages({
+  heading: {
+    defaultMessage: 'Time to start a new project?',
+    id: 'component.projects.new-project.heading'
+  },
+  lede: {
+    defaultMessage: 'Click the button to start a new project.',
+    id: 'component.projects.new-project.lede'
+  },
+  cta: {
+    defaultMessage: 'New project',
+    id: 'component.projects.new-project.cta'
+  }
+})
 
 interface IProjectsProps {
   projects: IProject[];
@@ -18,6 +35,8 @@ const Projects: React.FC<IProjectsProps> = ({
   includeCreateNew,
   includeBelonger
 }) => {
+  const history = useHistory()
+
   const { teamsById } = React.useContext(TeamsContext)
 
   const [search, setSearch] = React.useState<string>('')
@@ -40,7 +59,14 @@ const Projects: React.FC<IProjectsProps> = ({
         </>
       )}
 
-      {includeCreateNew && <NewProject />}
+      {includeCreateNew && (
+        <Empty
+          heading={messages.heading}
+          lede={messages.lede}
+          cta={messages.cta}
+          onClick={() => history.push('/new-project')}
+        />
+      )}
     </div>
   )
 }
